fix: refactor burger-ingredients component
