+++ src/components/burger-ingredients/burger-ingredients.module.css
@@ -6,88 +6,10 @@
     position: relative;
 }
 
-.ingredients__container {
+.ingredients {
     width: 100%;
     height: calc(100% - 200px);
     overflow-y: auto;
     overflow-x: hidden;
 }
 
-.ingredients {
-    width: 100%;
-}
-
-.ingredients__navigation {
-    display: flex;
-}
-
-.ingredients__list {
-    width: 100%;
-    display: grid;
-    margin: 0;
-    padding: 0;
-    grid-template-columns: repeat(auto-fit, 272px);
-    grid-auto-flow: row dense;
-    justify-content: center;
-    align-items: center;
-    list-style: none;
-    column-gap: 24px;
-    row-gap: 32px;
-}
-
-.ingredient {
-    width: 272px;
-    height: 208px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: flex-start;
-    position: relative;
-    transition: opacity .3s ease-in-out;
-}
-
-/*.ingredient:hover {*/
-/*    cursor: pointer;*/
-/*    opacity: .7;*/
-/*}*/
-
-.ingredient:hover > .ingredient__image {
-    cursor: pointer;
-    opacity: .6;
-}
-
-.ingredient:hover > .ingredient__price {
-    cursor: pointer;
-    opacity: .6;
-}
-
-.ingredient__counter {
-    position: absolute;
-    top: 0;
-    right: 0;
-}
-
-.ingredient__image {
-    width: 100%;
-    height: 120px;
-    object-fit: cover;
-    box-sizing: border-box;
-}
-
-.ingredient__price {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    justify-content: center;
-    column-gap: 10px;
-}
-
-.ingredient_text {
-    width: 100%;
-    margin: 0;
-    text-align: center;
-}
-
-.ingredient_text:hover {
-    cursor: pointer;
-}
