+++ src/components/burger-ingredients/ingredient/ingredient.module.css
@@ -1,0 +1,56 @@
+.container {
+    width: 272px;
+    height: 208px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: flex-start;
+    position: relative;
+    transition: opacity .3s ease-in-out;
+}
+
+/*.ingredient:hover {*/
+/*    cursor: pointer;*/
+/*    opacity: .7;*/
+/*}*/
+
+.container:hover > .ingredient__image {
+    cursor: pointer;
+    opacity: .6;
+}
+
+.container:hover > .ingredient__price {
+    cursor: pointer;
+    opacity: .6;
+}
+
+.counter {
+    position: absolute;
+    top: 0;
+    right: 0;
+}
+
+.image {
+    width: 100%;
+    height: 120px;
+    object-fit: cover;
+    box-sizing: border-box;
+}
+
+.price {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    column-gap: 10px;
+}
+
+.text {
+    width: 100%;
+    margin: 0;
+    text-align: center;
+}
+
+.text:hover {
+    cursor: pointer;
+}
