+++ src/components/burger-ingredients/ingredients-navigation/ingredients-navigation.module.css
@@ -1,0 +1,3 @@
+.container {
+    display: flex;
+}
