+++ src/utils/types.js
@@ -1,5 +1,14 @@
 import PropTypes from "prop-types";
 
+export const IngredientFullType = PropTypes.shape({
+  _id: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
+  price: PropTypes.number.isRequired,
+  image_mobile: PropTypes.string.isRequired,
+  image: PropTypes.string.isRequired,
+  count: PropTypes.number,
+})
+
 export const ConstructorElementType = PropTypes.shape({
   _id: PropTypes.string.isRequired,
   name: PropTypes.string.isRequired,
